+++ src/app/[lang]/projects/page.tsx
@@ -1,7 +1,0 @@
-export default function Projects(){
-    return(
-        <div>
-            <h1>Hello word!</h1>
-        </div>
-    )
-}